requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
openai==1.46.0
PyYAML==6.0.2
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
